strip out ill-considered 'extra' arg support
I added support for 'extra' args (those after a `--`) to the arg
parsing script for use in the terraform command. However, after further
consideration, I realized this is not a good way to handle the
situation.

The terraform command having been rewritten to not use this
parsing script further reinforces the fact that it should not have been
added. This commit strips that out.