Add `flop` cli (#1)
* allow nginx nodeport ports to be assigned via vars

* initial commit of flop

* strip out ill-considered 'extra' arg support

I added support for 'extra' args (those after a `--`) to the arg
parsing script for use in the terraform command. However, after further
consideration, I realized this is not a good way to handle the
situation.

The terraform command having been rewritten to not use this
parsing script further reinforces the fact that it should not have been
added. This commit strips that out.

* add note to README re flop resource requirements

* curl tests should fail if not 200

* flop: skip parsing -- on kubectl and terraform

* flop exec commands

* flop work around bash 3 bug re BASH_SOURCE

* flop fix tr incompatibilty with binary stream

* flop fix bash 3 bug with single quote in heredoc

* flop remove now unused vars

* flop create tf state dir if not existing

---------

Co-authored-by: Matt Bialas <mbialas@element84.com>